+++ src/ResourceManager/Resources/Commands.ResourceManager/Cmdlets/Implementation/ManagedApplications/NewAzureManagedApplication.cs
@@ -126,7 +126,8 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.Implementation
 
                 var activity = string.Format("PUT {0}", managementUri.PathAndQuery);
                 var result = this.GetLongRunningOperationTracker(activityName: activity, isResourceCreateOrUpdate: true)
-                    .WaitOnOperation(operationResult: operationResult);
+                    .WaitOnOperation(operationResult: operationResult, resourceId: resourceId, apiVersion: apiVersion);
+
                 this.WriteObject(this.GetOutputObjects("ManagedApplicationId", JObject.Parse(result)), enumerateCollection: true);
             }
         }